+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * The `weight_by` argument to `rep_slice_sample()` can now be passed either as a vector of numeric weights or an unquoted column name in `.data` (#480).
 
+* Fix issue with spaces in variable names (#472)
+
 # infer v1.0.4
 
 * Fixed bug in p-value shading where shaded regions no longer correctly overlaid